requirements.txt
numpy
opencv-python
sklearn
tqdm
paho-mqtt
flask
tensorflow-gpu